requirements.txt
djitellopy
numpy
opencv-python
python-dotenv
google-genai
google-cloud-vision
matplotlib
